+++ app/src/test/java/com/majkeylab/scanit/GeminiClientTest.kt
@@ -1,4 +1,4 @@
-package cz.mates.skendopdf
+package com.majkeylab.scanit
 
 import java.io.ByteArrayInputStream
 import java.io.IOException