+++ app/src/test/java/com/majkeylab/scanit/PureLogicTest.kt
@@ -1,4 +1,4 @@
-package cz.mates.skendopdf
+package com.majkeylab.scanit
 
 import java.io.File
 import java.nio.file.Files